pyproject.toml
[tool.poetry]
name = "hw_27"
version = "0.1.0"
description = ""
authors = ["Vip-Artyom <snowboard-team@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1"
Pillow = "^9.4.0"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
python-dateutil = "^2.8.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
